EFF-002

+++ WinFormsApp2/Models/Calculation.cs
@@ -2,10 +2,10 @@
 {
     public int Id { get; set; }
     public int UserId { get; set; }
-    public double InitialInvestment { get; set; }
-    public double DiscountRate { get; set; }
-    public double InflationRate { get; set; }
-    public double TaxRate { get; set; }
+    public decimal InitialInvestment { get; set; }
+    public decimal DiscountRate { get; set; }
+    public decimal InflationRate { get; set; }
+    public decimal TaxRate { get; set; }
     public int PoliticalStabilityRating { get; set; }
-    public double NPV { get; set; }
+    public decimal NPV { get; set; }
 }
